requirements.txt
# Core dependencies
aiohttp>=3.9,<4.0
Pillow
truststore
python-dateutil

# Web GUI dependencies (required)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
jinja2>=3.1.2
